Implement DoIP/CAN build time decoupling
